Splitting loadCV.js into smaller functions
